@@ -1,5 +1,5 @@
 module github.com/dedalqq/omg.httpserver
 
-go 1.17
+go 1.19
 
 require github.com/pkg/errors v0.9.1
